+++ __tests__/cli.js
@@ -8,29 +8,24 @@ const execOpts = { cwd: target }
 
 const phenomic = "node ./node_modules/.bin/phenomic"
 
-// cannot be done for now
-// see the mess in configurator which use process.argv directly
-// and since build.js is spawned, we got too many args and cannot use
-// yargs.strict()
-// https://github.com/MoOx/phenomic/issues/363
-// test.cb("should throw if a CLI flag is NOT recognized", (t) => {
-//   const child = exec(
-//     `${ phenomic } start --open=false --lol`, execOpts,
-//     (err) => {
-//       if (err && !err.killed) {
-//         clearTimeout(timeout)
-//         t.ok(err.message.indexOf("Unknown argument") > -1)
-//         t.end()
-//       }
-//     }
-//   )
-//
-//   const timeout = setTimeout(() => {
-//     child.kill()
-//     t.fail()
-//     t.end()
-//   }, 1000)
-// })
+test.cb("should throw if a CLI flag is NOT recognized", (t) => {
+  const child = exec(
+    `${ phenomic } start --open=false --lol`, execOpts,
+    (err) => {
+      if (err && !err.killed) {
+        clearTimeout(timeout)
+        t.truthy(err.message.indexOf("Unknown argument") > -1)
+        t.end()
+      }
+    }
+  )
+
+  const timeout = setTimeout(() => {
+    child.kill()
+    t.fail()
+    t.end()
+  }, 1000)
+})
 
 test.cb("should NOT throw if a CLI flag is recognized", (t) => {
   const child = exec(